feat: Update tauri dependency to version 1 with additional features
This commit updates the tauri dependency in the Cargo.toml file to version 1. It also adds the "macos-private-api" and "api-all" features to the tauri dependency. This update will enable the use of macOS private APIs and provide access to all available APIs. This change was made to enhance the functionality and capabilities of the application.